Refactored `getValueBytesForKey()` and `getCurrentValueFromTransientRecord()` from `ActiveActiveStoreIngestionTask` into `StorePartitionDataReceiver`. 🧹
